requirements.txt
streamlit
pymupdf
sentence-transformers
faiss-cpu
transformers
accelerate
torch
python-dotenv
numpy
pandas
hf_xet
